// src/core/enums/user-role-of-mainlocation-gulid-room.enums.ts
export enum UserRoleOfMainLocationGuildRoom {
  MainLocationGuildRoomManager = 1,
  MainLocationSecretariat = 11,
  MainLocationTreasurer = 12,
  MainLocationGuildRoomViceManager = 1003,
  MainLocationGuildRoomExecutiveManager = 1004,
}



// [Description("مدیر اتاق اصناف کشور")]
// MainLocationGuildRoomManager = 1,
// ​
// [Description("دبیرخانه اتاق اصناف کشور")]
// MainLocationSecretariat = 11,
// ​
// [Description("خزانه دار اتاق اصناف کشور")]
// MainLocationTreasurer = 12,
// ​
// [Description("نایب رئیس اتاق اصناف کشور")]
// MainLocationGuildRoomViceManager = 1003,
// ​
// [Description("مدیر اجرایی اتاق اصناف کشور")]
// MainLocationGuildRoomExecutiveManager = 1004,